Cleaning up some auth tests
